requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==4.0.5
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
psycopg2-binary
firebase-admin
djangorestframework==3.13.1
python-dotenv==0.12.0
